requirements.txt
tensorflow==2.20.0
scikit-learn
numpy
matplotlib
seaborn
pillow
kaggle
opencv-python

# Flask API for deployment
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
